+++ farm-backend/src/main/java/com/example/farmbackend/controller/ProductController.java
@@ -7,9 +7,12 @@ import jakarta.annotation.security.RolesAllowed;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static com.example.farmbackend.controller.ControllerUtils.checkAdminRole;
 
 @AllArgsConstructor
 @RestController
@@ -17,34 +20,53 @@ import java.util.List;
 public class ProductController {
 
     private ProductService productService;
-    @RolesAllowed({"ADMIN"})
-    @PostMapping
-    public ResponseEntity<ProductDto> createProduct(@RequestBody ProductDto productDto){
 
+    @PostMapping
+    public ResponseEntity<ProductDto> createProduct(@RequestBody ProductDto productDto,
+                                                    Authentication authentication
+    ){
+        if(!checkAdminRole(authentication)){
+            return  ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         ProductDto savedProduct = productService.createProduct(productDto);
         return  new ResponseEntity<>(savedProduct, HttpStatus.CREATED);
     }
-    @RolesAllowed({"ADMIN"})
+
     @GetMapping("/{name}")
-    public ResponseEntity<ProductDto> getProductById(@PathVariable("name") String name){
+    public ResponseEntity<ProductDto> getProductById(@PathVariable("name") String name,
+                                                     Authentication authentication){
+        if(!checkAdminRole(authentication)){
+            return  ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(productService.getProductByName(name));
     }
-    @RolesAllowed({"ADMIN"})
+
     @GetMapping
-    public ResponseEntity<List<ProductDto>> getAllProductDto(){
+    public ResponseEntity<List<ProductDto>> getAllProductDto(Authentication authentication){
+        if(!checkAdminRole(authentication)){
+            return  ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         List<ProductDto> products = productService.getAllProducts();
         return ResponseEntity.ok(products);
     }
-    @RolesAllowed({"ADMIN"})
+
     @PutMapping("/{name}")
     public ResponseEntity<ProductDto> updateProduct(@PathVariable("name") String name,
-                                                      @RequestBody ProductDto updateProduct){
+                                                      @RequestBody ProductDto updateProduct,
+                                                    Authentication authentication){
+        if(!checkAdminRole(authentication)){
+            return  ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         ProductDto productDto = productService.updateProduct(name, updateProduct);
         return  ResponseEntity.ok(productDto);
     }
-    @RolesAllowed({"ADMIN"})
+
     @DeleteMapping("/{name}")
-    public ResponseEntity<String> deleteProduct(@PathVariable("name")  String name){
+    public ResponseEntity<String> deleteProduct(@PathVariable("name")  String name,
+                                                Authentication authentication){
+        if(!checkAdminRole(authentication)){
+            return  ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         productService.delete(name);
         return  ResponseEntity.ok("Product deleted successfully!");
     }
